Fix negative expire values for refresh tokens (#11990)
When the client session max lifetime is set shorter than the sso session
max lifetime, at the point where the client session gets terminated, the
refresh token endpoint returns 400. When the user then tries to login
again the SSO remembers that there is a session but the resulting
refresh token of the token request has a negative expire time.

Now the getRefreshExpiration method correctly uses the clientSession to
calculate the expire time for the refresh token.

Fixes #11990